sam2/requirements.txt
decord==0.6.0
huggingface_hub==0.27.1
hydra-core==1.3.2
iopath==0.1.10
matplotlib==3.10.0
mediapipe==0.10.20
numpy==2.2.2
omegaconf==2.3.0
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
pycocotools==2.0.8
setuptools==75.8.0
skimage==0.0
submitit==1.5.2
tensordict==0.6.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
